First commit of Glintv2
This commit contains a lot of extra asset files native to the default django
admin site that can be removed as glintV2 progresses.

The codebase in this commit contains the basic framework for a django webserver
running behind apache over SSL. The configuration is set to only accept requests
using HTTPS and a user must have a valid grid certificate to even connect.

The templates are only for proof of concept and will be changed extinsively
as functionality is developed.

The database schema is also quite incomplete and should see notable chagnes
with the next commit.

For security reasons I've removed all key files from this commit and they
will need to be configured manually for the service to run.