add ability to save repository information to a project
This commit contains the code required to make a new repository for a given project
and add/save the credentials required to make contact with keystone.

The password field could probably use some additional security despite only passing thru
the ssl tunnel. Currently this repo configuration data isn't used for anything but will
be the bases of aquiring a token from that cloud as development continues.