Add celery and redis framework to allow for asynch tasks
This commit  contains the basic frameworks to have celery workers take care of periodic and
and asychronous tasks not suited for django. These tasks report their results to the local
redis server where a data structure of the image lists are maintained. This structure holds
the information about where the images are and what their current state is. This should allow
pending transfers to be done by celery workers completely outside the scope of the django
web-server.

There is still lots of work to do here but it has been a long time since i've commited and
I'm afraid of losing the work I've done. Some final notes I've installed a redis-server and
celery/celery beats on my development machine and they will need to be built into the ansible
script and init scripts will have to be made for them in production as I am currently running
them directly from the consul for development.