add state framework and save functionality
This commit contains the logic required to save the matrix state and queue up some transfer or delete jobs.
There is currently no logic to actually run transfer or deletes but the state change is shown in the matrix.

On a save the matrix is cross referenced with the old one and parsed into transfer tasks that are added
to the queue. The image collection thread then checks this queue for pending tasks every time it runs
and triggers the change of state in the redis image dictionary and will later start an asychnrounous
celery task to actually run the delete or transfer.