Update system to be focused around image name and add duplicate img name detection
This commit has the major chagnes required to centre glint around a name centric view of images instead of UUIDs.
Since we now have much less control over all the properties of images glint isn't able to predict all the
properties the new image will have when uploaded to a new repo. These changes allow glint to fill in the blanks
once the images have been flagged for a transfer.

Another big feature that comes along side the name centric focus is the ability to detect when a single repo
has two images with the same name. In this case when attempting to view the matrix you will be redirected to
a page to resolve the collision. The detection and form exist in this commit but the logic to handle the final
post still needs to be implemented.

Once the post logic has been implemented the only outstanding basic features are handeling the state changes
and actually performing the deletes and transfers. The code to perform transfers and deletes is relatively
simple but again we have to be careful when proccessing the state changes of the images.