add Delete and transfer functionality
With this commit the dev version of glint is fully operational.
I am able to resolve name conflicts within a repo, delete, and transfer images across repos.

There is still a ton of cleanup work to do and many strings should be moved to a config file.
The web page transitions are still not finalized and often just produce some text feedback
for testing. This will need to be updated to provide a good user experience.

Note: celery is unable to accept an object as a parameter and this resulted in passing
authentication info directly to the asynchronous celery task.