Fix add repo bug
Fixed a bug that allowed you to add two copies of the same repo.

This commit also has some small UI changes including but not limited to:
- Remove 2nd password field from add repo page
- Add redirect after saving a new repo

Lastly this commit adds a new field in the user model to track which project
they were last active on such that they can be redirected there on a reconnect.

The logic for the last point has not yet been implemented but the model should
now faccilitate this feature.