Update ssl error handling
This commit updates glint to provide better feedback when trying to add
additional repos to an account. I've also updated the certificate authority
bundle and added it to the config folder. NOTE that these are only the CAs
and not the certificates required for the server which should not be stored
in github.

Functionality has been loosely tested with dAir, cc-west and beaver.