add unique worker scratch folders, fixes #2
This commit makes a celery worker create its own unique temp directory to put images in.
This will let multiple workers work on images with the same name without fear of conflict.

Also fixed a bug where when submitting a transfer or delete while there was one pending
caused an extra transfer request to be requested.