add manage repos page #9
This commit has a basic manage repos page which is the last piece
of the puzzle for the basic user use cases. With this commit the
system should be ready for a production beta.

The current manage repo page is pretty poor in terms of interface,
this should be improved in further updates.

Before further user interface updates, the next step should be
to ansibilize deployment.