add celery init scripts and config files
This is the first step towards an ansible deployment.

The celery workers are now able to be run as a service.
celery beat is able to run in the background with the init script
but I will need to find a way to run the service and the celery beat
script on boot.