fix log file permissions issue and add alias for projects
The true solution to the log file issue that was stopping the celery daemons from running is that
they run as the celery user but the log file was owned by apache. By changing the ownership of
the log file to apache:celery they are now both able to access the file (664 permissinos).

I have added an alias field to the form when you add a project to an account and the matrix
now uses this alias a display name which should make it a lot easier to tell what you are
looking at.