Fix image dictionary, adresses #15
This commit contains cross cutting code changes to use the repo alias
instead of the repo name. This will fix the problem when two tenants
have the same name on different clouds glint wont try to lump them together
as a single repo.

The remaining change to make here is to add feedback when adding a new repo
as the alias field is now enforced unique and there is no error message
telling users that if they fail to add a repo for such a reason.