Remove periodic task and replace it with a perminant while loop
This change should allow glint to run without using the celerybeat scheduler.
This should improve performance since there is no queue delay or job distribution
done by the scheduler.

Instead of a periodic task the perminant task is called when apache loads up
and instantly starts monitoring the images of the repos in the database.

These changes are untested

: