move image collection to permanent loop
This commit continues progress towards #4 by having a constantly running celery
worker that does the image collection instead of a periodic task.

This should improve responsiveness of the system and allow us to remove the
celerybeat dependancy/proccess from glintv2.

This is working in my dev enviroment but the big problem is that there is no
clean way to shut down the worker. Sending a term signal asks the worker to
shut down after its finished it's current task which it will never do since
it is a perminent loop. Reading the celery docs I'm not sure if there is an
easy fix for this. The best thing to do would be to have apache cancel this
task when it shuts down so that celery wouldnt even need to restart unless
there were code changes for it. This will need some more thought.