Fix filter on refresh, shorten refresh window to 3 seconds
This commit is part of a series working towards #4.
This commit saves the last filter that was used in the browser's local
storage then is used on refresh to apply the same filter. Previously
every time the page refreshed the filter reverted to the default.

With this change in place the experience using glint should be much better
as it will provide you with much faster feedback after you request a delete
or transfer. The last major step will be to provide a % of how far along
the transfer is.