fix unique Alias bug
With the former implementation a given alias could only be used once
across all acounts which isn't ideal. I've eliminated the database
constraint for alias to be unique and provided a check for
alias-account combos when attempting to add a new repo.
This will allow different accounts to use the same alias' without
glint confusing them and putting them in the same dictionary.