Fix image filter
There was a bug that when you chose a filter and then started
typing in a text filter it would add images that violated the
selected public/private filter until the page reload. These
changes fix that issue so now when looking at private or
public images only the text filter will only show images from
the respective category.