Add dormant state to fix #19
There is now a new config option for setting the length of the downtime
cycle. The interval sets the downtime in multiples of 5 seconds.
The default value in glint is 1 minute and the default in the config
file is 5 minutes. Once glint has entered the dormant state every 5
seconds it checks if there is any new transactions or if it recieved
the termination signal. If there are transactions glint kicks into
'high gear' where there is no downtime between collection cycles
until all the transactions have completed to provide the fastest response
time. After they have completed it goes back into a dormant state