add image collection logic for new/deleted repos
This set of changes allows glint to update quickly after a new
repo has been added or deleted. Before if you added a repo at the
begining of a downtime cycle you might have had to wait up to 5 min
for the new repos images to be populated into the matrix.