Fix image collection false terminations
Fixes #20 by running a subprocess that checks if httpd is running.
If it sees that httpd is not running it will cleanly terminate the
image collection task and await another job or shutdown.

These chagnes should make the image collection task more stable
and allow for the httpd service to be shut down and restarted
without confusing the celery workers.