add image conflict detection
This commit is the fist step towards automated image conflict resolution.
Currently- at image collection time glint will detected conflicts
and compile a list to share with the web interface such that users
will be made aware of conflicts.

The next step will be to provide some level of automation for conflicts
glint has the authority to deal with.