add logic to upload image from file
This commit has the code required to successfully upload an image to one of the repos
registered with glint. Currently there is only support for a direct upload but logic
for uploading from a url should follow shortly.

This code still needs a lot of error handling as a failure causes the image collection
thread to be locked out of 'dormant' mode. This can be accomplished by a simple try
except block that decrements the number of transactions. More error handling should
also be added regarding user feedback when something has gone wrong with the upload.

The uploaded image should also be registered with the image cache such that it can
be removed after all of the uploads have finished. Right now it may cause errors to
delete the image after upload since we may be uploading to multiple repos.

The image cache needs a lot of fleshing out to be used for transfers and a routine
to clean up old files still needs to be developped and put into place.