progress towards #22 and #23
This commit contains chagnes to intellegently cache the images being uploaded
and the task to clean up any duplicates or files that don't qualify for the
cache list.

The last missing piece of the cache is an expirey time on the files and to
change the transfer task to consult the cache and add to it when transfering
images.