implement cache functionality into image transfer task
This commit improves the image transfer performance by taking advantage of the images already
cached locally.

Currently the cache cleanup task only removes files that are not registered in the cache table
but should be updated to also remove items that are old and unused such that the diskspace
doesn't get bloated with images nobody is interested in.