fix auto refresh bug
If the page was redirected to the project details page from another, when the automatic
refresh happened it would sometimes return to the previous page. This was fixed by
adding a url to the automatic refresh. However, this was a problem because it would
fully refresh the page removing any chagnes to the matrix or filters you may have
typed. This change should stop the page from being redirected to a previous page
and still keep the changes from the active session on a refresh.