add upload image button and update upload from URL
This commit has a fully working upload from url function. However, there aren't
any checks in place to make sure you are passing in a valid image. Ideally users
will give a proper url to glint during the upload, but if not it will still upload
the garbage file if the 'raw' format is selected.

Some additional testing on all upload features would be good:
 -Bad file with right/wrong format
 -Good file with right/wrong format
 -url to non-image file with different file formats (raw will upload regardles)
 -url with akward image name (image name is exctracted using the last '/'