Refactor image page, fixes #25, implement hidden image list
This commit contains many cross cutting changes regarding the image matrix page.
The idea of public/private images has been abstracted away by making images
either available in glint or hidden. Every image can be made hidden or made to be shown.
As a default all non-private images are hidden.

To facilitate these changes a 'hidden' attribute has been added to the image dictionary.
Additionally, the queue_state_change utility function has been modified to be able
to change the hidden state and the image state (present/pending/awaiting delete).

The matrix page now has two matrix that are accessible by clicking one of the tabs
at the top of the page (where the public/private filter used to be). The Glint
matrix functionality remains unchanged and you simple check the boxes of the images
you want to delete or transfer and click save images.
To toggle the hidden status of images you need to access the other matrix by clicking
the hidden images tab. Here the hidden status of each image is given by the checkboxes
if you'd like to hide an image make sure the box is checked and click hide images.
Similarily if you'd like to show images that were previously hidden, uncheck the boxes
and click the hide images button.