add static height to conflict message to avoid page jitter
This effectively fixes the problem of the jitter.
The remaining problem is that the buttons seems to flash on a reload as the css gets re-applied
if it is unable to load the style sheets before the content.
This problem could theoretically be solved if we used inline styling. However, I'd like to find
a better solution that could be more easily reused if there is another auto-refreshing page.